requirements.txt
groq
gradio
pdfplumber
python-docx
sentence-transformers
langchain-groq
faiss_cpu
tqdm
numpy